(GP-2.2): Created a Hash Function (#2)
* Adds basic git directory creation

Creates the foundation for a simplified Git implementation by setting up the initial directory structure.

This allows to initialize the "git" and "git/objects" directories, which are essential for storing repository data.

* Initializes Git repository structure.

Creates the basic directory structure for a Git repository,
including the 'git' directory and its 'objects' and 'index'
subdirectories.  This sets up the foundation for storing
and managing version control data.

* Initializes a basic Git repository structure

Sets up the foundational directory structure for a simplified Git repository.
Creates the 'git' directory, 'objects' subdirectory, 'index', and 'HEAD' files.
Adds a confirmation message when the repository is created.

Ensures that the creation process is idempotent by checking if the
directories and files already exist before attempting to create them.

* Documents git repo creation process

Documents the process for creating a git repository, including the creation of the objects subfolder and two files.

* (GP-2.1): Program Can Initialize Repository

The file contained outdated information regarding the initial implementation, so I updated it.

* (GP-2.2): Created a Hash Function

Implements SHA-1 hashing for files to track content changes.

Includes helper function to convert byte arrays to hexadecimal strings
for representation.