(GP-2.3) ALL
* (GP-2.3): Created Blob File

Adds functionality to create blob files in the 'git/objects' directory based on file content. It reads the file, generates a hash, and writes the content to a file named after the hash if it doesn't already exist.

* (GP-2.3.1): Blob Tester

Implements blob creation based on file content
and includes thorough testing to verify correct blob generation.
This ensures that files are accurately converted and stored as blobs.
The testing includes creating sample files, hashing them,
creating blob files, and then comparing the original file content
to the content of the generated blob file. Also, added a new clean up method for the samples directory.