All of Milestone 2.4 (#4)
* (GP-2.3) ALL

* (GP-2.3): Created Blob File

Adds functionality to create blob files in the 'git/objects' directory based on file content. It reads the file, generates a hash, and writes the content to a file named after the hash if it doesn't already exist.

* (GP-2.3.1): Blob Tester

Implements blob creation based on file content
and includes thorough testing to verify correct blob generation.
This ensures that files are accurately converted and stored as blobs.
The testing includes creating sample files, hashing them,
creating blob files, and then comparing the original file content
to the content of the generated blob file. Also, added a new clean up method for the samples directory.

* (GP-2.4): Updating the Index File

Adds a method to update the index file with new object hashes and file names.

Updates the tester to create the index file correctly.

* (GP-2.4.1): Index tester

Implements a test case to verify the correctness of the indexing mechanism.

The test creates sample files, adds them to the index, and then checks if the index contains the expected entries.

* (GP-2.4.2): Robust Reset Tester

Enables execution of indexing tests and implements a mechanism to reset the repository to a clean state, allowing for repeated and reliable testing. The reset functionality clears the objects directory, index file, and samples directory.

* (GP-2.4): Updating the Index File

Adds a method to update the index file with new object hashes and file names.

Updates the tester to create the index file correctly.

* (GP-2.4.1): Index tester

Implements a test case to verify the correctness of the indexing mechanism.

The test creates sample files, adds them to the index, and then checks if the index contains the expected entries.

* (GP-2.4.2): Robust Reset Tester

Enables execution of indexing tests and implements a mechanism to reset the repository to a clean state, allowing for repeated and reliable testing. The reset functionality clears the objects directory, index file, and samples directory.