Implements tree creation from index (#6)
* (GP-3.1): Updating the Index Formatting

Modifies the index update logic to handle file modifications
by updating the hash if the file content has changed. Also,
stores the relative path of files in the index instead of
just the file name. This allows the system to correctly
identify and track changes to files within the repository.
Adds a test case to verify that file modifications are
correctly tracked in the index.

* (GP-3.2): Creating a Basic Tree

Adds functionality to generate tree objects from directories.
Recursively traverses directories, creating blob objects for files
and tree objects for subdirectories. Writes tree objects to the object store.

* Tree from Working List Submitted before class(8 am)

Adds functionality to build trees from a working list.

This commit introduces a mechanism to parse and normalize the index,
sort the working list, and recursively create trees based on the file paths.
It finds the leaf-most parent directories, creates tree objects for their
children, and updates the working list iteratively until the entire
directory structure is represented as trees.

* (GP-3.3): Creating a Tree from the Index

Ensures the code gracefully handles scenarios where the input string
does not contain the expected number of parts after splitting.

* Tree from Working List Submitted before class(8 am)

Adds functionality to build trees from a working list.

This commit introduces a mechanism to parse and normalize the index,
sort the working list, and recursively create trees based on the file paths.
It finds the leaf-most parent directories, creates tree objects for their
children, and updates the working list iteratively until the entire
directory structure is represented as trees.

* (GP-3.3): Creating a Tree from the Index

Ensures the code gracefully handles scenarios where the input string
does not contain the expected number of parts after splitting.

* Tested worktoTree and made sure functionality was working

Implements the `workToTree` method to generate a tree structure based on the current working list.
This involves creating temporary files for hashing, writing tree contents, and creating tree files in the "git/objects" directory.
The method now returns the hash of the created tree, allowing for further operations.

Also renames tester file.